Fix 64-bit VS2010 build
R=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/420033003

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@22722 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/ast-graph-builder.h
@@ -196,7 +196,8 @@ class AstGraphBuilder::Environment
   int parameters_count() const { return parameters_count_; }
   int locals_count() const { return locals_count_; }
   int stack_height() {
-    return values()->size() - parameters_count_ - locals_count_;
+    return static_cast<int>(values()->size()) - parameters_count_ -
+           locals_count_;
   }
 
   // Operations on parameter or local variables. The parameter indices are
@@ -241,12 +242,12 @@ class AstGraphBuilder::Environment
   // Direct mutations of the operand stack.
   void Poke(int depth, Node* node) {
     ASSERT(depth >= 0 && depth < stack_height());
-    int index = values()->size() - depth - 1;
+    int index = static_cast<int>(values()->size()) - depth - 1;
     values()->at(index) = node;
   }
   Node* Peek(int depth) {
     ASSERT(depth >= 0 && depth < stack_height());
-    int index = values()->size() - depth - 1;
+    int index = static_cast<int>(values()->size()) - depth - 1;
     return values()->at(index);
   }
   void Drop(int depth) {
